synchronize

+++ src/core/StelFileMgr.cpp
@@ -26,13 +26,15 @@
 #include <QString>
 #include <QDebug>
 #include <QStandardPaths>
+#include <QProcessEnvironment>
+#include <QtGlobal>
 
-#include <stdio.h>
+#include <cstdio>
 
 #ifdef Q_OS_WIN
-#include <windows.h>
+#include <Windows.h>
 #ifndef _SHOBJ_H
-	#include <shlobj.h>
+	#include <ShlObj.h>
 	#include <QLibrary>
 #endif
 #endif
@@ -60,6 +62,19 @@ void StelFileMgr::init()
 	userDir = QDir::homePath() + "/.stellarium";
 #endif
 
+#if QT_VERSION >= 0x050A00
+	if (qEnvironmentVariableIsSet("STEL_USERDIR"))
+	{
+		userDir=qEnvironmentVariable("STEL_USERDIR");
+	}
+#else
+	QByteArray userDirCand=qgetenv("STEL_USERDIR");
+	if (userDirCand.length()>0)
+	{
+		userDir=QString::fromLocal8Bit(userDirCand);
+	}
+#endif
+
 	if (!QFile(userDir).exists())
 	{
 		qWarning() << "User config directory does not exist: " << QDir::toNativeSeparators(userDir);
@@ -75,15 +90,15 @@ void StelFileMgr::init()
 
 	// OK, now we have the userDir set, add it to the search path
 	fileLocations.append(userDir);
-
 	
 	// Determine install data directory location
+	QProcessEnvironment env = QProcessEnvironment::systemEnvironment();
+	QString envRoot = env.value("STELLARIUM_DATA_ROOT", ".");
 
-	// If we are running from the build tree, we use the files from the current directory
-	if (QFileInfo(CHECK_FILE).exists())
+	if (QFileInfo(envRoot + QDir::separator() + QString(CHECK_FILE)).exists())
 	{
-		installDir = ".";
-	}
+		installDir = envRoot;
+	}	
 	else
 	{
 	#if defined(Q_OS_MAC)
@@ -102,7 +117,7 @@ void StelFileMgr::init()
 			ResourcesDir.cd(QString("Resources"));
 		}
 		QFileInfo installLocation(ResourcesDir.absolutePath());
-		QFileInfo checkFile(installLocation.filePath() + QString("/") + QString(CHECK_FILE));
+		QFileInfo checkFile(installLocation.filePath() + QDir::separator() + QString(CHECK_FILE));
 	#elif defined(Q_OS_WIN)		
 		QFileInfo installLocation(QCoreApplication::applicationDirPath());
 		QFileInfo checkFile(installLocation.filePath() + QDir::separator() + QString(CHECK_FILE));
@@ -113,7 +128,7 @@ void StelFileMgr::init()
 		QFileInfo checkFile(QFile::decodeName(INSTALL_DATADIR "/" CHECK_FILE));
 	#endif
 
-	#ifndef NDEBUG
+	#ifdef DEBUG
 		if (!checkFile.exists())
 		{	// for DEBUG use sources location 
 			QString debugDataPath = INSTALL_DATADIR_FOR_DEBUG;
@@ -147,10 +162,27 @@ void StelFileMgr::init()
 						 << QDir::toNativeSeparators(relativePath)
 						 << " (we checked for "
 						 << QDir::toNativeSeparators(checkFile.filePath()) << ").";
-				#ifndef UNIT_TEST
-				// NOTE: Hook for buildbots (using within testEphemeris)
-				qFatal("Couldn't find install directory location.");
-				#endif
+
+				qWarning() << "Maybe this is development environment? Let's check source directory path...";
+
+				QString sourceDirPath = STELLARIUM_SOURCE_DIR; // The variable is defined in CMakeLists.txt file
+				checkFile = QFileInfo(sourceDirPath + QDir::separator() + CHECK_FILE);
+				if (checkFile.exists())
+				{
+					installDir = sourceDirPath;
+				}
+				else
+				{
+					qWarning() << "WARNING StelFileMgr::StelFileMgr: could not find install location:"
+							 << QDir::toNativeSeparators(sourceDirPath)
+							 << " (we checked for "
+							 << QDir::toNativeSeparators(checkFile.filePath()) << ").";
+
+					#ifndef UNIT_TEST
+					// NOTE: Hook for buildbots (using within testEphemeris)
+					qFatal("Couldn't find install directory location.");
+					#endif
+				}
 			}
 		}
 	}
@@ -386,7 +418,6 @@ bool StelFileMgr::fileFlagsCheck(const QFileInfo& thePath, const Flags& flags)
 
 QString StelFileMgr::getDesktopDir()
 {
-
 	if (QStandardPaths::standardLocations(QStandardPaths::DesktopLocation).isEmpty())
 		return "";
 
@@ -495,12 +526,12 @@ QString StelFileMgr::getWin32SpecialDirPath(int csidlId)
 	// therefore it's using only the wide-char version of the code. --BM
 	QLibrary library(QLatin1String("shell32"));
 	typedef BOOL (WINAPI*GetSpecialFolderPath)(HWND, LPTSTR, int, BOOL);
-	GetSpecialFolderPath SHGetSpecialFolderPath = (GetSpecialFolderPath)library.resolve("SHGetSpecialFolderPathW");
+	GetSpecialFolderPath SHGetSpecialFolderPath = reinterpret_cast<GetSpecialFolderPath>(library.resolve("SHGetSpecialFolderPathW"));
 	if (SHGetSpecialFolderPath)
 	{
 		TCHAR tpath[MAX_PATH];
-		SHGetSpecialFolderPath(0, tpath, csidlId, FALSE);
-		return QString::fromUtf16((ushort*)tpath);
+		SHGetSpecialFolderPath(Q_NULLPTR, tpath, csidlId, FALSE);
+		return QString::fromUtf16(reinterpret_cast<ushort*>(tpath));
 	}
 
 	return QString();
